Replace ctrlp with FZF in vim
ctrlp is usually too slow to be useful, and FZF has more functionality.
As most of the FZF commands (namely :Files and :Rg) benefit from an
initial query string argument, the <C-p> binding now maps to :Buffers to
make it easier to actually use buffers instead of relying so heavily on
windows and tabs.